Reorganize bricks subpackage.
Splits up bricks/__init__.py into a few separate modules, with
__init__.py taking on the typical role of importing specific
symbols from each of them and exporting the blocks.bricks API.

This makes the code easier to navigate, and also easier to write
individual modules within blocks/bricks/ which depend on other
blocks.bricks functionality, without incurring import loops.

This also makes #885 trivial to implement for blocks.bricks: simply
import symbols from the rest of the .py files under blocks/bricks/
and add them to __all__ (which is necessary for flake8 not to complain,
and defines the list of things imported by 'import *', and is generally
a nice-to-have).
